Moved shared fixtures to application folder.
